+++ apps/web/app/(pages)/create-profile3/page.tsx
@@ -1,4 +1,4 @@
-import Home from "@/components/Feature/ProfileLayout/Layout2/Home";
+import Home from "@/components/Feature/ProfileLayout/Layout3/Home";
 
 export default function CreateProfilePage() {
   return <Home />;
