--- apps/web/components/Feature/CreateProfile1Home/index.tsx
@@ -1,12 +1,0 @@
-"use client";
-
-import CreateProfile1Layout from "../../Layout/CreateProfile1";
-import CreateProfile1Hero from "../CreateProfile1Hero";
-
-export default function CreateProfile1Home() {
-  return (
-    <CreateProfile1Layout>
-      <CreateProfile1Hero />
-    </CreateProfile1Layout>
-  );
-}
