--- apps/web/components/Feature/CreateProfile2About/index.tsx
@@ -1,12 +1,0 @@
-"use client";
-
-import CreateProfileLayout from "../../Layout/CreateProfile";
-import CreateProfileHeroFrame from "../CreateProfile2Hero";
-
-export default function CreateProfileAbout() {
-  return (
-    <CreateProfileLayout>
-      <CreateProfileHeroFrame />
-    </CreateProfileLayout>
-  );
-}
