--- apps/web/components/Feature/CreateProfile2Collections/index.tsx
@@ -1,14 +1,0 @@
-"use client";
-
-import CreateProfileLayout from "../../Layout/CreateProfile";
-import CreateProfileHeroFrame from "../CreateProfile2Hero";
-import CollectionList from "./ColletionList";
-
-export default function CreateProfileCollections() {
-  return (
-    <CreateProfileLayout>
-      <CreateProfileHeroFrame />
-      <CollectionList />
-    </CreateProfileLayout>
-  );
-}
