+++ apps/web/components/Feature/ProfileLayout/Layout1/Collections/index.tsx
@@ -1,0 +1,14 @@
+"use client";
+
+import CreateProfile1Layout from "@/components/Layout/CreateProfile1";
+import CollectionList from "../../Layout2/Collections/ColletionList";
+import Hero from "../Hero";
+
+export default function Collections() {
+  return (
+    <CreateProfile1Layout>
+      <Hero />
+      <CollectionList />
+    </CreateProfile1Layout>
+  );
+}
