+++ apps/web/components/Feature/ProfileLayout/Layout1/Home/index.tsx
@@ -1,0 +1,12 @@
+"use client";
+
+import CreateProfile1Layout from "@/components/Layout/CreateProfile1";
+import Hero from "../Hero";
+
+export default function CreateProfile1Home() {
+  return (
+    <CreateProfile1Layout>
+      <Hero />
+    </CreateProfile1Layout>
+  );
+}
