+++ apps/web/components/Feature/ProfileLayout/Layout2/About/index.tsx
@@ -1,0 +1,12 @@
+"use client";
+
+import CreateProfileLayout from "../../../../Layout/CreateProfile2";
+import Hero from "../Hero";
+
+export default function About() {
+  return (
+    <CreateProfileLayout>
+      <Hero />
+    </CreateProfileLayout>
+  );
+}
