+++ apps/web/components/Feature/ProfileLayout/Layout2/Collections/index.tsx
@@ -1,0 +1,14 @@
+"use client";
+
+import CreateProfileLayout from "../../../../Layout/CreateProfile2";
+import Hero from "../Hero";
+import CollectionList from "./ColletionList";
+
+export default function Collections() {
+  return (
+    <CreateProfileLayout>
+      <Hero />
+      <CollectionList />
+    </CreateProfileLayout>
+  );
+}
