+++ apps/web/components/Feature/ProfileLayout/Layout2/Home/index.tsx
@@ -1,0 +1,16 @@
+"use client";
+
+import CreateProfileLayout from "../../../../Layout/CreateProfile2";
+import LastestUpload from "./LastestUpload";
+import Hero from "../Hero";
+import AboutSection from "../About/AboutSection";
+
+export default function Home() {
+  return (
+    <CreateProfileLayout>
+      <Hero />
+      <LastestUpload />
+      <AboutSection />
+    </CreateProfileLayout>
+  );
+}
