+++ apps/web/components/Feature/ProfileLayout/Layout3/Home/index.tsx
@@ -1,0 +1,12 @@
+"use client";
+
+import CreateProfile1Layout from "@/components/Layout/CreateProfile1";
+import Hero from "../Hero";
+
+export default function Home() {
+  return (
+    <CreateProfile1Layout>
+      <Hero />
+    </CreateProfile1Layout>
+  );
+}
